requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt setup.py
#
apispec[marshmallow]==6.6.1
    # via flask-smorest
blinker==1.8.2
    # via flask
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
flask==3.0.3
    # via
    #   flask-smorest
    #   flask_opentelemetry_app (setup.py)
flask-smorest==0.44.0
    # via flask_opentelemetry_app (setup.py)
googleapis-common-protos==1.63.2
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.66.0
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-otlp-proto-grpc
idna==3.8
    # via requests
importlib-metadata==8.0.0
    # via
    #   opentelemetry-api
    #   opentelemetry-instrumentation-flask
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.22.0
    # via
    #   apispec
    #   flask-smorest
    #   webargs
opentelemetry-api==1.26.0
    # via
    #   flask_opentelemetry_app (setup.py)
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-jaeger==1.11.1
    # via flask_opentelemetry_app (setup.py)
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift==1.11.1
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp==1.26.0
    # via flask_opentelemetry_app (setup.py)
opentelemetry-exporter-otlp-proto-common==1.26.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.26.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.26.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.47b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-flask==0.47b0
    # via flask_opentelemetry_app (setup.py)
opentelemetry-instrumentation-requests==0.47b0
    # via flask_opentelemetry_app (setup.py)
opentelemetry-instrumentation-wsgi==0.47b0
    # via opentelemetry-instrumentation-flask
opentelemetry-proto==1.26.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.26.0
    # via
    #   flask_opentelemetry_app (setup.py)
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.47b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.47b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
packaging==24.1
    # via
    #   apispec
    #   marshmallow
    #   opentelemetry-instrumentation-flask
    #   webargs
protobuf==4.25.4
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
requests==2.32.3
    # via opentelemetry-exporter-otlp-proto-http
six==1.16.0
    # via thrift
thrift==0.20.0
    # via opentelemetry-exporter-jaeger-thrift
typing-extensions==4.12.2
    # via opentelemetry-sdk
urllib3==2.2.2
    # via requests
webargs==8.4.0
    # via flask-smorest
werkzeug==3.0.4
    # via
    #   flask
    #   flask-smorest
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.20.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
